Extracted feed sync from DownloadService
