+++ core/src/main/java/de/danoeh/antennapod/core/service/download/handler/FailedDownloadHandler.java
@@ -1,0 +1,33 @@
+package de.danoeh.antennapod.core.service.download.handler;
+
+import android.util.Log;
+import de.danoeh.antennapod.core.feed.Feed;
+import de.danoeh.antennapod.core.service.download.DownloadRequest;
+import de.danoeh.antennapod.core.service.download.DownloadStatus;
+import de.danoeh.antennapod.core.storage.DBWriter;
+
+/**
+ * Handles failed downloads.
+ * <p/>
+ * If the file has been partially downloaded, this handler will set the file_url of the FeedFile to the location
+ * of the downloaded file.
+ * <p/>
+ * Currently, this handler only handles FeedMedia objects, because Feeds and FeedImages are deleted if the download fails.
+ */
+public class FailedDownloadHandler implements Runnable {
+    private static final String TAG = "FailedDownloadHandler";
+    private final DownloadRequest request;
+
+    public FailedDownloadHandler(DownloadRequest request) {
+        this.request = request;
+    }
+
+    @Override
+    public void run() {
+        if (request.getFeedfileType() == Feed.FEEDFILETYPE_FEED) {
+            DBWriter.setFeedLastUpdateFailed(request.getFeedfileId(), true);
+        } else if (request.isDeleteOnFailure()) {
+            Log.d(TAG, "Ignoring failed download, deleteOnFailure=true");
+        }
+    }
+}
