+++ core/src/main/java/de/danoeh/antennapod/core/storage/DBReader.java
@@ -515,13 +515,24 @@ public final class DBReader {
      * newest events first.
      */
     public static List<DownloadStatus> getFeedDownloadLog(Feed feed) {
-        Log.d(TAG, "getFeedDownloadLog() called with: " + "feed = [" + feed + "]");
+        return getFeedDownloadLog(feed.getId());
+    }
+
+    /**
+     * Loads the download log for a particular feed from the database.
+     *
+     * @param feedId Feed id for which the download log is loaded
+     * @return A list with DownloadStatus objects that represent the feed's download log,
+     * newest events first.
+     */
+    public static List<DownloadStatus> getFeedDownloadLog(long feedId) {
+        Log.d(TAG, "getFeedDownloadLog() called with: " + "feed = [" + feedId + "]");
 
         PodDBAdapter adapter = PodDBAdapter.getInstance();
         adapter.open();
         Cursor cursor = null;
         try {
-            cursor = adapter.getDownloadLog(Feed.FEEDFILETYPE_FEED, feed.getId());
+            cursor = adapter.getDownloadLog(Feed.FEEDFILETYPE_FEED, feedId);
             List<DownloadStatus> downloadLog = new ArrayList<>(cursor.getCount());
             while (cursor.moveToNext()) {
                 downloadLog.add(DownloadStatus.fromCursor(cursor));
